[refactor] : separate ApiService class
